Separated out the two viewmodels for the home page and app page
If a user is logged in they get the AppViewModel. If not logged in,  then the user gets the HomeViewModel. AppViewModel uses the user's data (i.e. remaining requests), while the HomeViewModel uses an unregistered user's remaining calls based off IP address. Not a perfect solution because a user can change their IP and bypass the need for registering. An unreg. user has their IP stored in the DB and a registered user does not.