Notification bug-fixes. Code optimization and cleanup. New design features.
Fixed an issue with the notifications that are displayed to the user when their email is confirmed. Fixed an issue that prevented the loading screen from covering the entire page, when the app is reloaded.  Broader exception handling in the tokenizer script (renamed from email_token). New fonts added. New icons for the footer links (github, email). Restyled some templates with different colors, and fonts.