Tests, general code refactoring.
I wrote tests for the completion service, makes use of pytest and fixtures.Fixed an issue with db_session that prevented some tests from using the in-memory db. The check to see if the db is already initialized (Global __factory has a value) has been moved to the configuration that takes place within the __init__.py. Removed the flask app factory pattern i.e. no longer defining a function called create_app, and instead am just running the configuration directly in the __init__ .  This allows the app to be imported easily for testing. Also refactored the app and home views to make use of extensive error handling, should the openai api not work as expected.